Add timer to provide live progress bar of audio

+++ src/main/java/com/zischase/discordbot/audioplayer/PlayerPrinter.java
@@ -6,6 +6,7 @@ import com.sedmelluq.discord.lavaplayer.player.event.TrackExceptionEvent;
 import com.sedmelluq.discord.lavaplayer.player.event.TrackStuckEvent;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
+import com.sun.istack.Nullable;
 import com.zischase.discordbot.DBQueryHandler;
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.MessageBuilder;
@@ -14,63 +15,78 @@ import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.entities.TextChannel;
 
 import java.awt.*;
+import java.time.Instant;
 import java.time.OffsetDateTime;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 public class PlayerPrinter {
 
-	public PlayerPrinter(AudioManager audioManager, TextChannel defaultChannel) {
+	private final int   historyPollLimit = 25;
+	private final int   npTimerRateMs    = 3000;
+	private final Timer nowPlayingTimer  = new Timer("nowPlayingTimer");
 
+	public PlayerPrinter(AudioManager audioManager, TextChannel defaultChannel) {
 		AudioEventListener trackWatcherEventListener = audioEvent -> {
+			/* Check for available channel to display Now PLaying prompt */
 			String      dbQuery       = DBQueryHandler.get(defaultChannel.getGuild().getId(), "media_settings", "textChannel");
 			TextChannel activeChannel = defaultChannel.getGuild().getTextChannelById(dbQuery);
 
+			/* Clear the current timer, we got a new event to handle */
+			nowPlayingTimer.purge();
+
+			/* Ensure we have somewhere to send the message, check for errors */
 			assert activeChannel != null;
 			if (audioEvent instanceof TrackStuckEvent) {
 				activeChannel.sendMessage("Audio track stuck!").queue();
+				audioEvent.player.stopTrack();
 			} else if (audioEvent instanceof TrackExceptionEvent) {
 				activeChannel.sendMessage("Error loading the audio.").queue();
-			} else {
+			} else if (audioEvent.player.getPlayingTrack() == null) {
 				printNowPlaying(audioManager, activeChannel);
+				activeChannel.getJDA().getDirectAudioController().disconnect(activeChannel.getGuild());
+			} else {
+				/* Set up a timer to continually update the running time of the song */
+				nowPlayingTimer.scheduleAtFixedRate(new TimerTask() {
+					@Override
+					public void run() {
+						if (audioEvent.player.getPlayingTrack() != null) {
+							if (audioEvent.player.getPlayingTrack().getPosition() < audioEvent.player.getPlayingTrack().getDuration()) {
+								printNowPlaying(audioManager, activeChannel);
+							}
+						}
+					}
+				}, Date.from(Instant.now()), npTimerRateMs);
 			}
 		};
 
+		/* Add the event watcher to the current guild's audio manager */
 		audioManager.getPlayer().addListener(trackWatcherEventListener);
 	}
 
 	public void printNowPlaying(AudioManager audioManager, TextChannel channel) {
-		AudioPlayer player = audioManager.getPlayer();
-		deletePrevious(channel, "Now Playing");
+		AudioPlayer  player = audioManager.getPlayer();
+		EmbedBuilder embed  = new EmbedBuilder();
 
-		EmbedBuilder embed = new EmbedBuilder();
-		embed.setColor(Color.CYAN);
-
+		long autoDeleteDelayMS;
 		if (player.getPlayingTrack() == null) {
 			embed.setTitle("Nothing Playing");
 			embed.setColor(Color.darkGray);
 			embed.setFooter(". . .");
+			autoDeleteDelayMS = 5000;
 		} else {
-			AudioTrackInfo info = player.getPlayingTrack().getInfo();
-
-
-			long duration = info.length / 1000;
-			long position = player.getPlayingTrack().getPosition() / 1000;
-
-			String timeTotal = String.format("%d:%02d:%02d", duration / 3600, (duration % 3600) / 60, (duration % 60));
-
+			AudioTrackInfo info      = player.getPlayingTrack().getInfo();
+			long           duration  = info.length / 1000;
+			long           position  = player.getPlayingTrack().getPosition() / 1000;
+			String         repeat    = audioManager.getScheduler().isRepeat() ? " \uD83D\uDD01" : "Off";
+			String         timeTotal = String.format("%d:%02d:%02d", duration / 3600, (duration % 3600) / 60, (duration % 60));
 			String timeCurrent = String.format("%d:%02d:%02d",
 					position / 3600,
 					(position % 3600) / 60,
 					(position % 60)
 			);
-
-
-//			embed.setThumbnail(Config.get("MEDIA_PLAYER_ICON"));
-			embed.setTitle("Now Playing");
 
 			if (info.title != null && !info.title.isEmpty()) {
 				embed.appendDescription(info.title + "\n\n");
@@ -80,49 +96,62 @@ public class PlayerPrinter {
 				embed.appendDescription("-----\n\n");
 			}
 
-
 			if (player.isPaused()) {
-				embed.appendDescription("Paused");
+				embed.appendDescription("Paused -  \uD83D\uDD34 ");
 			} else if (player.getPlayingTrack().getDuration() == Long.MAX_VALUE) {
-				embed.appendDescription("Live");
+				embed.appendDescription("Live - \uD83C\uDF99");
 			} else {
-				embed.appendDescription(timeCurrent + " - " + timeTotal);
+				embed.appendDescription(timeCurrent + " - " + timeTotal + "");
 				String progressBar = progressPercentage((int) position, (int) duration);
 				embed.appendDescription(System.lineSeparator() + progressBar);
 			}
-
+			embed.setColor(Color.CYAN);
+			embed.setTitle("\uD83D\uDCFB \uD83C\uDFB6 Now Playing \uD83C\uDFB6 \uD83D\uDCFB");
+			embed.appendDescription("\n"+repeat);
 			embed.setFooter(info.uri);
+			autoDeleteDelayMS = player.getPlayingTrack().getDuration() - player.getPlayingTrack().getPosition();
 		}
 
-		long delayMS = 2000;
-		if (player.getPlayingTrack() != null) {
-			delayMS = player.getPlayingTrack().getDuration() - player.getPlayingTrack().getPosition();
+		Message message = new MessageBuilder().setEmbeds(embed.build()).build();
+		Message currentMsg = getCurrentNowPlayingMsg(channel);
+
+		if (currentMsg != null) {
+			channel.editMessageById(currentMsg.getId(), message).queue();
 		}
 
-		Message message = new MessageBuilder().setEmbeds(embed.build())
-				.build();
-
-		long finalDelayMS = delayMS;
-
-
-		channel.sendMessage(message)
-				.queue(msg ->
-				{
-					if (channel.getHistory().getRetrievedHistory().contains(msg)) {
-						msg.delete()
-								.queueAfter(finalDelayMS, TimeUnit.MILLISECONDS);
-					}
-				});
+		else {
+			channel.sendMessage(message)
+					.queue(msg ->
+					{
+						if (channel.getHistory().getRetrievedHistory().contains(msg)) {
+							msg.delete()
+									.queueAfter(autoDeleteDelayMS, TimeUnit.MILLISECONDS);
+						}
+					});
+		}
 	}
 
-	private void deletePrevious(TextChannel textChannel, String titleSearch) {
+	@Nullable
+	private Message getCurrentNowPlayingMsg(TextChannel textChannel) {
+		return textChannel
+				.getHistory()
+				.retrievePast(historyPollLimit)
+				.complete()
+				.stream()
+				.filter(msg -> msg.getAuthor().isBot())
+				.filter(msg -> !msg.getEmbeds().isEmpty())
+				.filter(msg -> !msg.isPinned())
+				.filter(msg -> msg.getTimeCreated().isBefore(OffsetDateTime.now()))
+				.filter(msg -> msg.getTimeCreated().isBefore(OffsetDateTime.now().plusDays(14)))
+				.filter(msg -> Objects.requireNonNull(msg.getEmbeds().get(0).getTitle()).contains("Now Playing"))
+				.findFirst()
+				.orElse(null);
+	}
 
-		if (titleSearch == null) {
-			return;
-		}
+	private void deletePrevious(TextChannel textChannel) {
 
 		List<Message> msgList = textChannel.getHistory()
-				.retrievePast(25)
+				.retrievePast(historyPollLimit)
 				.complete()
 				.stream()
 				.filter(msg ->
@@ -139,31 +168,33 @@ public class PlayerPrinter {
 					if (isBot && hasEmbeds && notPinned && createdBeforeNow && lessThan2Weeks) {
 						MessageEmbed embed = msg.getEmbeds().get(0);
 						if (embed.getTitle() != null) {
-							return embed.getTitle()
-									.matches("(?i)(" + titleSearch + "|Nothing Playing)");
+							return embed.getTitle().contains("Now Playing") || embed.getTitle().contains("Queue") || embed.getTitle().contains("Nothing Playing");
 						}
 					}
 					return false;
 				})
 				.collect(Collectors.toList());
 
+		/* Using blocking method to preserve execution order in class - Print queue attempts to rewrite a message that gets deleted here before success */
 		if (msgList.size() > 1) {
-			textChannel.deleteMessages(msgList).queue();
+			textChannel.deleteMessages(msgList).complete();
+		} else if (msgList.size() == 1) {
+			textChannel.deleteMessageById(msgList.get(0).getId()).complete();
 		}
 	}
 
 	private String progressPercentage(int done, int total) {
-		int    size              = 10;
+		int    size              = 30;
 		String iconLeftBoundary  = "|";
-		String iconDone          = "-";
-		String iconRemain        = "\u00A0"; // &nbsp
+		String iconDone          = "=";
+		String iconRemain        = "\u00A0. \u200b";
 		String iconRightBoundary = "|";
 
 		if (done > total) {
 			throw new IllegalArgumentException();
 		}
 		int donePercent = (100 * done) / total;
-		int doneLength   = size * donePercent / 100;
+		int doneLength  = (size * donePercent) / 100;
 
 		StringBuilder bar = new StringBuilder(iconLeftBoundary);
 		for (int i = 0; i < size; i++) {
@@ -187,19 +218,11 @@ public class PlayerPrinter {
 		embed.setColor(Color.BLUE);
 		embed.setTitle("Queue");
 
-		deletePrevious(channel, "Queue");
+		deletePrevious(channel);
 
 		if (queue.isEmpty()) {
 			embed.appendDescription("Nothing in the queue.");
-
-			channel.sendMessageEmbeds(embed.build())
-					.queue(msg ->
-					{
-						if (channel.getHistory().getRetrievedHistory().contains(msg)) {
-							msg.delete()
-									.queueAfter(5000, TimeUnit.MILLISECONDS);
-						}
-					});
+			channel.sendMessageEmbeds(embed.build()).complete();
 			return;
 		}
 
@@ -237,18 +260,7 @@ public class PlayerPrinter {
 
 		embed.appendDescription(" ```fix\nUp Next: " + queue.get(queue.size() - 1).getInfo().title + "```");
 
-		channel.sendMessageEmbeds(embed.build())
-				.queue(msg ->
-				{
-					try {
-						Thread.sleep(5000);
-					} catch (InterruptedException e) {
-						e.printStackTrace();
-					}
-					if (channel.getHistory().getRetrievedHistory().contains(msg)) {
-						msg.delete().complete();
-					}
-				});
+		channel.sendMessageEmbeds(embed.build()).complete();
 	}
 
 }
